third_party/WebKit/LayoutTests/fast/dom/shadow/shadow-root-direction-expected.txt

CONSOLE WARNING: line 8: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Element direction details are accessible to shadow root

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(shadowRoot.getElementById("target2")).direction is "rtl"
PASS getComputedStyle(root).direction is "rtl"
PASS successfullyParsed is true

TEST COMPLETE
